+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- add PrometheusExporter::Ext::Server::ProcStatCollector [#8](https://github.com/didww/prometheus_exporter-ext/pull/8)
 
 ## [0.2.3] - 2024-05-03
 - add PrometheusExporter::Ext::Instrumentation::ProcStat [#7](https://github.com/didww/prometheus_exporter-ext/pull/7)